- Review comments processed Related to Issue [WTI-434]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@2606 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: f05f341be86ca1b4eb879333768a3925aa0cf58f
Former-commit-id: 447cb6e97ffc42fd36f3f0ce3d1649cd844161b9

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/ReferenceLinePointEntity.cs
@@ -39,7 +39,7 @@ namespace Application.Ringtoets.Storage.DbContext
         public Nullable<long> DikeAssessmentSectionEntityId { get; set; }
         public Nullable<decimal> X { get; set; }
         public Nullable<decimal> Y { get; set; }
-        public Nullable<long> Order { get; set; }
+        public Nullable<int> Order { get; set; }
     
         public virtual DikeAssessmentSectionEntity DikeAssessmentSectionEntity { get; set; }
     }
